Hoist locking giant back up into the ioctl handler
Move the locking back into the ioctl handler. This "fixes" the race where we hve
a hot plug event just after the dropping of Giant in pci_find_dbsf, assuming the
driver doesn't then call anything that drops and picks up Giant again... It's a
little safer since don't think it doesn't, but we lack the tools to know for
sure.


git-svn-id: https://svn.freebsd.org/base/head@355057 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f