Remove SPARE_USRSPACE.
This constant was used to reserve space at the top of the stack to
hold translated system call arguments for non-default ABIs (the
"stackgap").  However, none of the compatibility ABIs have used the
stackgap in many years and the last use of SPARE_USRSPACE was removed
in r355373.

Reviewed by:	kib
Sponsored by:	DARPA
Differential Revision:	https://reviews.freebsd.org/D22697


git-svn-id: https://svn.freebsd.org/base/head@355459 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f