arm64: rk3328: Add the *clk_peri_niu clocks
Those clocks are always enable by default and are not really explained
in the TRM but the reason we had them is that they have the periph clock
as a parent and those parent should never be disable which can happen
if we disable all the childs. The current childs are the sd/emmc/sdio clocks
so the board will hang if we disable them.

MFC after:	1 month


git-svn-id: https://svn.freebsd.org/base/head@355624 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f