MFC r355568:
Correctly check for C++17 and higher when declaring timespec_get()

Summary:
In rS338751, the check to declare `timespec_get()` for C++17 and higher
was incorrectly done against a `cplusplus` define, while it should have
been `__cplusplus`.

Fix this by using `__cplusplus`, and also bump `__FreeBSD_version` so it
becomes possible to correctly check for `timespec_get()` in upstream
libc++ headers.

Reviewed by:	brooks, emaste
Differential Revision: https://reviews.freebsd.org/D22735


git-svn-id: https://svn.freebsd.org/base/stable/12@355658 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f