vt: store a pointer to the keyboard instead of index
This effectively reverts r355935, but is functionally equivalent. We gain no
benefit from storing the index and repeatedly fetching the keyboard with
`kbd_get_keyboard` when we need it. We'll be notified when it's going away
so we can clean up the pointer.

All existing references were trivially converted. Only once instance
actually needed the index.


git-svn-id: https://svn.freebsd.org/base/head@355937 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f