Fix kbdd_diag definition
This is a direct commit to stable branches as these have been converted to
inline functions in head, which will likely not get MFC'd due to limited
utility -- follow up locking assertions likely cannot be done in stable
branches.

The use of 'level' in the macro definition was truncated to 'leve', which
makes the macro unusable.


git-svn-id: https://svn.freebsd.org/base/stable/12@356011 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f