[PowerPC] Switch to PIC kernel modules on powerpc*
To improve reliability of kernel modules after the clang switch, switch to
-fPIC when building for now.

This bypasses some limitations to the way clang and LLD handle relocations,
and is a more robustly tested compilation regime than the
"static shared object" mode that we were previously attempting to convince
the compiler stack to use.

The kernel linker was recently augmented to be able to handle this mode.

Reviewed by:	jhibbits
Differential Revision:	https://reviews.freebsd.org/D22798


git-svn-id: https://svn.freebsd.org/base/head@356114 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f