inetd: convert remaining bzero(3) to memset(3), NFC
This change is purely in the name of noise reduction from static analyzers
that want to complain that bzero(3) is obsolete in favor of memset(3).

With this, clang-analyze at least is now noise free. WARNS= 6 also appears
to have been OK for some time now, so drop the current setting and opt for
the default.


git-svn-id: https://svn.freebsd.org/base/head@356248 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f